Bye android helpers + nasty classpath change
As it was the web app libraries took precedence over the JRE - and in those
there is the android.jar - _and in there java io 5/6 classes_. So it would
NOT see the Java 7 ones. Nasty.

Fixed in Java build path > _Order_ and export and reflected in .classpath.
So the change in the order of <classpathentry> IS IMPORTANT

+ debug prints for file IO (delete them)